Remove start button from quiz challenge
Having to think about how to delay the fetch is an added layer of
complexity that doesn't add much value. We could maybe re-add "Have they
player press start before starting" as an extra optional thing at the
end, but it's making harder for some of our students to get started at
all.